add cursorRadius field to CamState and refactor cursorLocation
cursorLocation no longer requires an explicit radius parameter, as it
uses the radius built into the CamState record. The upshot of this is
that we can now easily adjust the cursor radius in a settings file, for
instance.

This required refactoring cursorLocation, which previously took a radius
as a parameter. However, advanceCamera used cursorLocation with a small,
explicit radius to determine where to move the camera to when the
movement keys are pressed down. To deal with that, we created
cursorLocation', allowing for an explicit radius, cursorLocation was
rewritten to use cursorLocation'.